requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
python-decouple==3.6
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
Pillow==9.2.0
flake8==5.0.4
django-cors-headers==3.13.0

#schema
pyyaml==6.0
uritemplate==4.1.1
coreapi==2.3.3